# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartWarehouse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type RegisterRequestRole =
  (typeof RegisterRequestRole)[keyof typeof RegisterRequestRole];

export const RegisterRequestRole = {
  buyer: "buyer",
  dealer: "dealer",
  admin: "admin",
} as const;

export interface RegisterRequest {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterRequestRole;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  buyer: "buyer",
  dealer: "dealer",
  admin: "admin",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export type CreateOrderRequestMaterial =
  (typeof CreateOrderRequestMaterial)[keyof typeof CreateOrderRequestMaterial];

export const CreateOrderRequestMaterial = {
  cement: "cement",
  steel: "steel",
  sand: "sand",
} as const;

export interface CreateOrderRequest {
  material: CreateOrderRequestMaterial;
  totalQty: number;
  location: string;
  deliveryDate: string;
  notes?: string;
}

export type OrderMaterial = (typeof OrderMaterial)[keyof typeof OrderMaterial];

export const OrderMaterial = {
  cement: "cement",
  steel: "steel",
  sand: "sand",
} as const;

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  requested: "requested",
  quoted: "quoted",
  allocated: "allocated",
  dispatched: "dispatched",
  delivered: "delivered",
} as const;

export interface Order {
  id: string;
  buyerId: string;
  buyerName?: string;
  material: OrderMaterial;
  totalQty: number;
  location: string;
  deliveryDate: string;
  notes?: string;
  status: OrderStatus;
  createdAt: string;
}

export interface CreateQuotationRequest {
  orderId: string;
  pricePerUnit: number;
  availableQty: number;
  deliveryDate: string;
}

export interface Quotation {
  id: string;
  orderId: string;
  dealerId: string;
  dealerName?: string;
  pricePerUnit: number;
  availableQty: number;
  deliveryDate: string;
  createdAt: string;
}

export type SubOrderStatus =
  (typeof SubOrderStatus)[keyof typeof SubOrderStatus];

export const SubOrderStatus = {
  allocated: "allocated",
  dispatched: "dispatched",
  delivered: "delivered",
} as const;

export interface SubOrder {
  id: string;
  orderId: string;
  dealerId: string;
  dealerName?: string;
  allocatedQty: number;
  pricePerUnit: number;
  status: SubOrderStatus;
  material?: string;
  location?: string;
  createdAt: string;
}

export type UpdateStatusRequestStatus =
  (typeof UpdateStatusRequestStatus)[keyof typeof UpdateStatusRequestStatus];

export const UpdateStatusRequestStatus = {
  dispatched: "dispatched",
  delivered: "delivered",
} as const;

export interface UpdateStatusRequest {
  status: UpdateStatusRequestStatus;
}