# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartWarehouse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestMaterial } from "./createOrderRequestMaterial";

export interface CreateOrderRequest {
  material: CreateOrderRequestMaterial;
  totalQty: number;
  location: string;
  deliveryDate: string;
  notes?: string;
}