# lib/api-zod/src/generated/types/createQuotationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartWarehouse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuotationRequest {
  orderId: string;
  pricePerUnit: number;
  availableQty: number;
  deliveryDate: string;
}