# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartWarehouse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderMaterial } from "./orderMaterial";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  buyerId: string;
  buyerName?: string;
  material: OrderMaterial;
  totalQty: number;
  location: string;
  deliveryDate: string;
  notes?: string;
  status: OrderStatus;
  createdAt: string;
}